Automatically backed up by Learn

+++ lib/create.sql
@@ -2,4 +2,4 @@ CREATE TABLE projects (id INTEGER PRIMARY KEY, title TEXT, category TEXT, fundin
 
 CREATE TABLE users (id INTEGER PRIMARY KEY, name TEXT, age INTEGER);
 
-CREATE TABLE pledges (id INTEGER PRIMARY KEY, amount FLOAT, project_id INTEGER, user_id INTEGER);
+CREATE TABLE pledges (id INTEGER PRIMARY KEY, amount INTEGER, project_id INTEGER, user_id INTEGER);
